# Modified Authorizations.of to handle duplicate input elements