--- a/src/test/java/org/apache/accumulo/access/AccessEvaluatorTest.java
+++ b/src/test/java/org/apache/accumulo/access/AccessEvaluatorTest.java
@@ -84,14 +84,14 @@ public class AccessEvaluatorTest {
       AccessEvaluator evaluator;
       assertTrue(testSet.auths.length >= 1);
       if (testSet.auths.length == 1) {
-        evaluator = AccessEvaluator.of(testSet.auths[0]);
+        evaluator = AccessEvaluator.of(Authorizations.of(Set.of(testSet.auths[0])));
         runTestCases(testSet, evaluator);
 
         Set<String> auths = Stream.of(testSet.auths[0]).collect(Collectors.toSet());
         evaluator = AccessEvaluator.of(auths::contains);
         runTestCases(testSet, evaluator);
       } else {
-        var authSets = Stream.of(testSet.auths).map(a -> Authorizations.of(false, a))
+        var authSets = Stream.of(testSet.auths).map(a -> Authorizations.of(Set.of(a)))
             .collect(Collectors.toList());
         evaluator = AccessEvaluator.of(authSets);
         runTestCases(testSet, evaluator);
@@ -184,11 +184,14 @@ public class AccessEvaluatorTest {
 
   @Test
   public void testEmptyAuthorizations() {
-    assertThrows(IllegalArgumentException.class, () -> AccessEvaluator.of(""));
-    assertThrows(IllegalArgumentException.class, () -> AccessEvaluator.of("", "A"));
-    assertThrows(IllegalArgumentException.class, () -> AccessEvaluator.of("A", ""));
     assertThrows(IllegalArgumentException.class,
-        () -> AccessEvaluator.of(Authorizations.of(false, "")));
+        () -> AccessEvaluator.of(Authorizations.of(Set.of(""))));
+    assertThrows(IllegalArgumentException.class,
+        () -> AccessEvaluator.of(Authorizations.of(Set.of("", "A"))));
+    assertThrows(IllegalArgumentException.class,
+        () -> AccessEvaluator.of(Authorizations.of(Set.of("A", ""))));
+    assertThrows(IllegalArgumentException.class,
+        () -> AccessEvaluator.of(Authorizations.of(Set.of(""))));
   }
 
   @Test