@dlmarion 2024-03-12
> I don't really like the boolean, but these change fix the problem of the inconsistent behavior and document it and test it. Personally I think we should drop the boolean and pick one consistent behavior (either dedupe or throw an exception) for both methods and document the behavior in javadoc and test it.

I don't like the boolean or throwing the exception, but this is middle-ground. Personally, I think we should just change the API to accept a Set instead of a Collection or array. That removes all ambiguity and places the burden on the user to supply us with a unique set of values without throwing an exception.