@keith-turner 2024-03-12
> Personally, I think we should just change the API to accept a Set instead of a Collection or array. That removes all ambiguity and places the burden on the user to supply us with a unique set of values without throwing an exception.

I like a single method that takes a set.  I saw that comment on #67 and meant to respond to, but did not immediately respond and then forgot about it.  There is only one thing I don't like about the  single set method is that it makes writing examples a bit more verbose, but losing that is not a bit deal.  I don't think the varargs `of(String ... auhs)` method has much practical use that I can think of, like I doubt it would ever be used outside of example or unit test code.  So having the single method that takes a set seems like a good overall solution.