@keith-turner 2024-03-12
We may want to consider removing [this code](https://github.com/apache/accumulo-access/blob/2da31d3643916c020b14c9e61fa632543dbc56b7/src/main/java/org/apache/accumulo/access/AccessEvaluator.java#L155-L160) also.  I have no idea what its behavior is for dupes.   Its not documented and I did not look at the impl.  This method also does not have much practical use other than making example code less verbose.