@dlmarion 2024-03-12
I ran the benchmark and compared to main the changes were in the noise.

Main:
```

Benchmark                                           Mode  Cnt   Score   Error   Units
AccessExpressionBenchmark.measureBytesValidation   thrpt   12  33.087 ± 0.139  ops/us
AccessExpressionBenchmark.measureEvaluation        thrpt   12  12.863 ± 0.109  ops/us
AccessExpressionBenchmark.measureStringValidation  thrpt   12  25.500 ± 0.144  ops/us
```
This branch:
```
Benchmark                                           Mode  Cnt   Score   Error   Units
AccessExpressionBenchmark.measureBytesValidation   thrpt   12  33.022 ± 0.171  ops/us
AccessExpressionBenchmark.measureEvaluation        thrpt   12  12.949 ± 0.355  ops/us
AccessExpressionBenchmark.measureStringValidation  thrpt   12  25.577 ± 0.314  ops/us
```